♻️ Refactor saveRequirementToArtifact from tool to workflow node
Convert saveRequirementToArtifactTool from PMAnalysisAgent tool to dedicated workflow node following invokeSchemaDesignToolNode pattern.

Key changes:
- Remove saveRequirementToArtifactTool from PMAnalysisAgent.bindTools()
- Create saveRequirementToArtifactNode as separate workflow termination node
- Update workflow routing: analyzeRequirements → saveRequirementToArtifact → dbAgent
- Add conditional edge retry logic for failed requirements analysis
- Update tests and documentation with new workflow diagram
- Use WorkflowTerminationError for consistent error handling
