+++ frontend/internal-packages/agent/src/chat/workflow/nodes/analyzeRequirementsNode.ts
@@ -1,90 +1,10 @@
-import { AIMessage } from '@langchain/core/messages'
 import type { RunnableConfig } from '@langchain/core/runnables'
 import type { Database } from '@liam-hq/db'
 import { PMAnalysisAgent } from '../../../langchain/agents'
-import type { Repositories } from '../../../repositories'
 import { WorkflowTerminationError } from '../../../shared/errorHandling'
 import { getConfigurable } from '../shared/getConfigurable'
 import type { WorkflowState } from '../types'
 import { logAssistantMessage } from '../utils/timelineLogger'
-import {
-  createOrUpdateArtifact,
-  transformWorkflowStateToArtifact,
-} from '../utils/transformWorkflowStateToArtifact'
-import { withTimelineItemSync } from '../utils/withTimelineItemSync'
-
-/**
- * Format analyzed requirements into a structured string
- */
-const formatAnalyzedRequirements = (
-  analyzedRequirements: NonNullable<WorkflowState['analyzedRequirements']>,
-): string => {
-  const formatRequirements = (
-    requirements: Record<string, string[]>,
-    title: string,
-  ): string => {
-    const entries = Object.entries(requirements)
-    if (entries.length === 0) return ''
-
-    return `${title}:
-${entries
-  .map(
-    ([category, items]) =>
-      `- ${category}:\n  ${items.map((item) => `  • ${item}`).join('\n')}`,
-  )
-  .join('\n')}`
-  }
-
-  const sections = [
-    `Business Requirement:\n${analyzedRequirements.businessRequirement}`,
-    formatRequirements(
-      analyzedRequirements.functionalRequirements,
-      'Functional Requirements',
-    ),
-    formatRequirements(
-      analyzedRequirements.nonFunctionalRequirements,
-      'Non-Functional Requirements',
-    ),
-  ].filter(Boolean)
-
-  return sections.join('\n\n')
-}
-
-/**
- * Save artifacts if workflow state contains artifact data
- */
-async function saveArtifacts(
-  state: WorkflowState,
-  repositories: Repositories,
-  assistantRole: Database['public']['Enums']['assistant_role_enum'],
-): Promise<void> {
-  if (!state.analyzedRequirements && !state.generatedUsecases) {
-    return
-  }
-
-  const artifact = transformWorkflowStateToArtifact(state)
-  const artifactResult = await createOrUpdateArtifact(
-    state,
-    artifact,
-    repositories,
-  )
-
-  if (artifactResult.success) {
-    await logAssistantMessage(
-      state,
-      repositories,
-      'Your requirements have been analyzed and saved',
-      assistantRole,
-    )
-  } else {
-    await logAssistantMessage(
-      state,
-      repositories,
-      'Unable to save your requirements. Please try again or contact support...',
-      assistantRole,
-    )
-  }
-}
 
 /**
  * Analyze Requirements Node - Requirements Organization
@@ -117,13 +37,6 @@ export async function analyzeRequirementsNode(
 
   return analysisResult.match(
     async (analysisData) => {
-      const analyzedRequirements = {
-        businessRequirement: analysisData.response.businessRequirement,
-        functionalRequirements: analysisData.response.functionalRequirements,
-        nonFunctionalRequirements:
-          analysisData.response.nonFunctionalRequirements,
-      }
-
       // Log reasoning summary if available
       if (
         analysisData.reasoning?.summary &&
@@ -139,31 +52,19 @@ export async function analyzeRequirementsNode(
         }
       }
 
-      // Create complete message with all analyzed requirements and sync to timeline
-      const completeMessage = await withTimelineItemSync(
-        new AIMessage({
-          content: formatAnalyzedRequirements(analyzedRequirements),
-          name: 'PMAnalysisAgent',
-        }),
-        {
-          designSessionId: state.designSessionId,
-          organizationId: state.organizationId || '',
-          userId: state.userId,
-          repositories,
-          assistantRole,
-        },
-      )
-
-      const updatedState = {
-        ...state,
-        messages: [completeMessage],
-        analyzedRequirements,
+      // PMAnalysisAgent has already parsed the JSON response
+      // Set analyzedRequirements so the conditional routing can proceed to saveRequirementToArtifact
+      const analyzedRequirements = {
+        businessRequirement: analysisData.response.businessRequirement,
+        functionalRequirements: analysisData.response.functionalRequirements,
+        nonFunctionalRequirements:
+          analysisData.response.nonFunctionalRequirements,
       }
 
-      // Save artifacts if requirements are successfully analyzed
-      await saveArtifacts(updatedState, repositories, assistantRole)
-
-      return updatedState
+      return {
+        ...state,
+        analyzedRequirements,
+      }
     },
     async (error) => {
       await logAssistantMessage(
